pyproject.toml
[tool.poetry]
name = "aospdtgen"
version = "0.1.0"
description = "Create an Android device tree from a device dump"
authors = ["Sebastiano Barezzi <barezzisebastiano@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
pyelftools = "^0.27"
Jinja2 = "^3.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
